Move buffer and image limits out of topology handling
These have no place in pocl_topology, and are now set in their own
function, implemented by basic and used by both basic and pthread.

Also, fix a missing check against the OpenCL-commanded minimum of 128M
and exclude getrlimit() call under Windows.
