+++ lib/CL/devices/basic/basic.c
@@ -36,6 +36,7 @@
 
 #ifndef _MSC_VER
 #  include <sys/time.h>
+#  include <sys/resource.h>
 #  include <unistd.h>
 #else
 #  include "vccompat.hpp"
@@ -330,6 +331,72 @@ pocl_basic_probe(struct pocl_device_ops *ops)
   return env_count;
 }
 
+#define MIN_MAX_MEM_ALLOC_SIZE (128*1024*1024)
+
+/* set maximum allocation sizes for buffers and images */
+void
+pocl_basic_set_buffer_image_limits(cl_device_id device)
+{
+  /* Maximum allocation size: we don't have hardware limits, so we
+   * can potentially allocate the whole memory for a single buffer, unless
+   * of course there are limits set at the operating system level. Of course
+   * we still have to respect the OpenCL-commanded minimum */
+  size_t alloc_limit = SIZE_MAX;
+
+#ifndef _MSC_VER
+  // TODO getrlimit equivalent under Windows
+  struct rlimit limits;
+  int ret = getrlimit(RLIMIT_DATA, &limits);
+  if (ret == 0)
+    alloc_limit = limits.rlim_cur;
+#endif
+  if (alloc_limit > device->global_mem_size)
+    alloc_limit = device->global_mem_size;
+  else if (alloc_limit < MIN_MAX_MEM_ALLOC_SIZE)
+    alloc_limit = MIN_MAX_MEM_ALLOC_SIZE;
+  // TODO in theory now if alloc_limit was > rlim_cur and < rlim_max
+  // we should try and setrlimit to alloc_limit, or allocations might fail
+
+  device->local_mem_size = device->max_constant_buffer_size =
+    device->max_mem_alloc_size = alloc_limit;
+
+  /* We don't have hardware limitations on the buffer-backed image sizes,
+   * so we set the maximum size in terms of the maximum amount of pixels
+   * that fix in max_mem_alloc_size. A single pixel can take up to 4 32-bit channels,
+   * i.e. 16 bytes.
+   */
+  size_t max_pixels = device->max_mem_alloc_size/16;
+  if (max_pixels > device->image_max_buffer_size)
+    device->image_max_buffer_size = max_pixels;
+
+  /* Similarly, we can take the 2D image size limit to be the largest power of 2
+   * whose square fits in image_max_buffer_size; since the 2D image size limit
+   * starts at a power of 2, it's a simple matter of doubling.
+   * This is actually completely arbitrary, another equally valid option
+   * would be to have each maximum dimension match the image_max_buffer_size.
+   */
+  max_pixels = device->image2d_max_width;
+  // keep doubing until we go over
+  while (max_pixels <= device->image_max_buffer_size/max_pixels)
+    max_pixels *= 2;
+  // halve before assignment
+  max_pixels /= 2;
+  if (max_pixels > device->image2d_max_width)
+    device->image2d_max_width = device->image2d_max_height = max_pixels;
+
+  /* Same thing for 3D images, of course with cubes. Again, totally arbitrary. */
+  max_pixels = device->image3d_max_width;
+  // keep doubing until we go over
+  while (max_pixels*max_pixels <= device->image_max_buffer_size/max_pixels)
+    max_pixels *= 2;
+  // halve before assignment
+  max_pixels /= 2;
+  if (max_pixels > device->image3d_max_width)
+  device->image3d_max_width = device->image3d_max_height =
+    device->image3d_max_depth = max_pixels;
+
+}
+
 void
 pocl_basic_init (cl_device_id device, const char* parameters)
 {
@@ -351,6 +418,7 @@ pocl_basic_init (cl_device_id device, const char* parameters)
   device->data = d;
   pocl_topology_detect_device_info(device);
   pocl_cpuinfo_detect_device_info(device);
+  pocl_basic_set_buffer_image_limits(device);
 
   /* The basic driver represents only one "compute unit" as
      it doesn't exploit multiple hardware threads. Multiple
