+++ lib/CL/devices/pthread/pthread.c
@@ -261,6 +261,7 @@ pocl_pthread_init (cl_device_id device, const char* parameters)
 
   pocl_topology_detect_device_info(device);
   pocl_cpuinfo_detect_device_info(device);
+  pocl_basic_set_buffer_image_limits(device);
 
   if(!strcmp(device->llvm_cpu, "(unknown)"))
     device->llvm_cpu = NULL;
